==> globalprep/spp/v2023/_setup/2_set_up_iucn_risk_regional.Rmd <==
---
title: 'Set up IUCN marine species risk by regional assessment'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '../../../../workflow/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---

```{r}
knitr::opts_chunk$set(echo = TRUE, eval=FALSE)
```


``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'Figs/',
                      echo = TRUE, message = FALSE, warning = FALSE, eval = FALSE)

```


```{r}
library(dplyr)
library(rgdal)
library(raster)
library(here)
library(ggridges)
library(ggplot2)

source(here('workflow/R/common.R'))

### provenance tracking
#library(provRmd); prov_setup()

goal     <- 'spp'
scenario <- 'v2022'
dir_goal <- here('globalprep', goal, scenario)

### goal specific folders and info
dir_anx  <- file.path(dir_M, 'git-annex/globalprep')
dir_setup   <- file.path(dir_goal, '_setup')
dir_goal_anx <- file.path(dir_anx, goal, scenario, 'spp_risk_dists')

source(file.path(dir_setup, 'common_fxns.R'))

```

# Summary

Examine IUCN Red List regional assessments for conservation status categories.

# Data Sources

### IUCN Red List

# Methods

## Get IUCN regional assessments

### Get list of regional assessments

Pull the IDs and species counts for all the regional IUCN Red List assessments.

``` {r get_rgn_assess_info_from_api}
### Get all pages and bind into total species list.  This is pretty fast.

rgn_assess_from_api_file <- file.path(dir_setup, sprintf('int/spp_rgn_assess_%s.csv', api_version))

message('Using API to get IUCN regional assessment info')

rgn_assess_url <- sprintf('http://apiv3.iucnredlist.org/api/v3/region/list?token=%s', api_key)
rgn_assess <- fromJSON(rgn_assess_url)$results
n_assess <- nrow(rgn_assess)

spp_count <- vector(length = n_assess, 'integer')
for (i in 1:n_assess) { ### i <- 1
  rgn_spp_count_url <- sprintf('http://apiv3.iucnredlist.org/api/v3/speciescount/region/%s?token=%s',
                               rgn_assess$identifier[i], api_key)
  spp_count[i] <- fromJSON(rgn_spp_count_url)$count
}
  
rgn_assess$spp_count <- spp_count
rgn_assess$version   <- api_version

rgn_assess <- rgn_assess %>%
  filter(identifier != 'global') 

write_csv(rgn_assess, rgn_assess_from_api_file)

```
  
### For each regional assessment, get species assessment info

Regional Assessments:

  
For each assessment, read the pages of species info.  Pages are 10k spp long, so break into individual pages and then attach.  Species lists will be matched to the marine habitats list, and updated category codes and scores attached.

```{r, eval = FALSE}

rgn_assess <- read_csv(rgn_assess_from_api_file)

rgn_spp_info_file <- file.path(dir_goal, sprintf('_data/iucn_risk_rgn_current_%s.csv', api_version))

if(!file.exists(rgn_spp_info_file)) {
  
  spp_df_rgn_list <- vector(length = nrow(rgn_assess), 'list')
  for(i in 1:nrow(rgn_assess)) {  ###i <- 1
    rgn_id <- rgn_assess$identifier[i]
    n_pages <- ceiling(rgn_assess$spp_count[i] / 10000)
    spp_page_url <- file.path(sprintf('http://apiv3.iucnredlist.org/api/v3/species/region/%s', rgn_id),
                              'page/%s?token=%s')
      ### fill in the region identifier, leave the other two as %s for mc_get_from_api
    
    spp_df_rgn <- mc_get_from_api(spp_page_url, c(0:(n_pages - 1)), api_key, delay = 1)
    spp_df_rgn <- spp_df_rgn %>%
      dplyr::select(-infra_rank, -infra_name, -count, -page) %>%
      setNames(str_replace(tolower(names(.)), '_name', '')) %>%
      rename(iucn_rgn = region_identifier, sciname = scientific) %>%
      mutate(api_version = api_version)
    
    spp_df_rgn_list[[i]] <- spp_df_rgn
  }
  
  spp_df_rgn_all <- bind_rows(spp_df_rgn_list)
  
  ### filter to just marine species, and select relevant columns
  spp_rgn_marine <- read_csv(file.path(dir_goal, '_data',
                                   sprintf('spp_marine_from_api_%s.csv', api_version))) %>%
    inner_join(spp_df_rgn_all, by = c('iucn_sid' = 'taxonid')) %>%
    dplyr::select(iucn_sid, sciname,
           old_cat = category, iucn_rgn)
  
  cat_lookup <- read_csv(file.path(dir_setup, 'raw', 'risk_code_lookup.csv')) %>%
    dplyr::select(cat = code, cat_score = cat_score, cat_current = code_current) %>%
    filter(!is.na(cat))
  
  spp_rgn_marine1 <- spp_rgn_marine %>%
    left_join(cat_lookup, by = c('old_cat' = 'cat'))


  write_csv(spp_rgn_marine1, rgn_spp_info_file)
  
} else {
  
  message('File of API species regional assessments exists: \n  ', rgn_spp_info_file)
  git_prov(rgn_spp_info_file, filetype = 'output')
  
}

```

Once regional assessments are gleaned from the API, check them against the global marine species list.  Species are matched by ID numbers.  Species whose regional category differs from the global category are kept and scored.

* note, quick check of scientific names showed they all match
* 113 spp fall into this...

``` {r}

global_marine_spp_risk <- read_csv(file.path(dir_goal, sprintf('_data/iucn_risk_current_%s.csv', api_version)))

rgn_marine_spp <- read_csv(rgn_spp_info_file)

rgn_spp_list <- global_marine_spp_risk %>%
  dplyr::select(iucn_sid, sciname, main_common_name, global_cat = cat, global_cat_score = cat_score) %>%
  inner_join(rgn_marine_spp %>%
               dplyr::select(iucn_sid, 
                       #rgn_sciname = sciname, #rgn_pop = population,
                      rgn_cat = cat_current, rgn_cat_score = cat_score, iucn_rgn),
             by = 'iucn_sid') %>%
  filter(global_cat_score != rgn_cat_score)

```

## Get IUCN regional historical assessments

``` {r get_iucn_past_assessments}

spp_timeseries_file <- file.path(dir_goal, '_data',
                                 sprintf('iucn_risk_ts_rgn_%s.csv', api_version))

global_marine_spp_risk <- read_csv(file.path(dir_goal, sprintf('_data/iucn_risk_current_%s.csv', api_version)))

rgn_marine_spp <- read_csv(rgn_spp_info_file)

cat_lookup <- read_csv(file.path(dir_setup, 'raw', 'risk_code_lookup.csv')) %>%
  dplyr::select(rgn_cat = code, rgn_cat_ts_score = cat_score, rgn_cat_ts = code_current) %>%
  filter(!is.na(rgn_cat))

rgn_assess <- read_csv(rgn_assess_from_api_file)

if(!file.exists(spp_timeseries_file)) {
  
  spp_ts_df_rgn_list <- vector(length = nrow(rgn_assess), 'list')
  for(i in 1:nrow(rgn_assess)) {  #i <- 9
    rgn_id <- rgn_assess$identifier[i]
    
    spp_hist_url <- file.path('http://apiv3.iucnredlist.org/api/v3/species', 
                                  'history/id/%s/region/RGN_ID?token=%s') %>%
      str_replace('RGN_ID', rgn_id)
      ### fill in the region identifier, leave the other two as %s for mc_get_from_api
    
    spp_rgn_ids <- rgn_marine_spp %>%
      filter(iucn_rgn == rgn_id) %>%
      .$iucn_sid %>%
      unique()
    
    spp_ts_df_rgn <- mc_get_from_api(spp_hist_url, spp_rgn_ids, api_key, delay = 1)
    
    if(names(spp_ts_df_rgn) != c('param_id', 'api_error')) {
      spp_ts_df_rgn <- spp_ts_df_rgn %>%
        dplyr::select(iucn_sid = name, year, rgn_cat = code, iucn_rgn = region_identifier) %>%
        filter(!is.na(iucn_sid))
    }
    
    spp_ts_df_rgn <- spp_ts_df_rgn %>%
      mutate(api_version = api_version)
    
    spp_ts_df_rgn_list[[i]] <- spp_ts_df_rgn
  }
  
  spp_ts_df_rgn_all <- bind_rows(spp_ts_df_rgn_list)

  
  spp_hist_df1 <- spp_ts_df_rgn_all %>% 
    left_join(cat_lookup, by = c('rgn_cat' = 'rgn_cat')) %>%
    filter(!is.na(rgn_cat_ts_score) & !is.na(year)) %>%
    arrange(iucn_sid, year) %>%
    dplyr::select(iucn_sid, year, rgn_cat_ts, rgn_cat_ts_score, iucn_rgn) %>%
    mutate(iucn_version = api_version) %>%
    group_by(iucn_sid, iucn_rgn) %>%
    #summarise(n())
    filter(n() >= 2)
  
  write_csv(spp_hist_df1, spp_timeseries_file)
  
} else {
  
  git_prov(spp_timeseries_file, filetype = 'output')
  
}

```

-----

``` {r prov_footer, results = 'asis'}

prov_wrapup(commit_outputs = FALSE)

```
